src/corp_speech_risk_dataset/corpus_temporal_cv/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "corpus-temporal-cv"
version = "1.0.0"
description = "Temporal cross-validation module for corpus dataset"
authors = [
    {name = "Corpus Team", email = "corpus@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer>=0.4.0",
    "loguru>=0.5.0",
    "pydantic>=1.8.0",
    "pyyaml>=5.4.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scikit-learn>=1.0.0",
    "torch>=1.9.0",
    "joblib>=1.0.0",
]

[project.scripts]
corpus-cv-run = "corpus_temporal_cv.cli.cv:run"
corpus-cv-report = "corpus_temporal_cv.cli.cv:report"

[tool.setuptools.packages.find]
where = ["src"]
include = ["corpus_temporal_cv*"]
